整合胡建飞的项目 3月29日 解决了项目本地生成图片的路径问题 解决了nexus私服问题
