+++ backend-manager/backend-manager-dao/src/main/java/cn/yutang/backend/dao/DinnerTableInfoMapper.java
@@ -1,9 +1,15 @@
 package cn.yutang.backend.dao;
 
+import cn.yutang.backend.pojo.dto.Page;
 import cn.yutang.backend.pojo.vo.DinnerTableInfo;
+import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 
 public interface DinnerTableInfoMapper {
-    List<DinnerTableInfo> selectByTtId(Integer ttId);
+    List<DinnerTableInfo> selectByPage(@Param("ttId") Integer ttId, @Param("page") Page page);
+
+    List<DinnerTableInfo> listAll();
+
+    int selectCountByTtId(Integer ttId);
 }
