+++ backend-manager/backend-manager-dao/src/main/java/cn/yutang/backend/dao/DinnerTableInfoMapper.xml
@@ -9,7 +9,7 @@
         <id column="tb_id" jdbcType="INTEGER" property="tbId" />
         <result column="tb_name" jdbcType="VARCHAR" property="tbName" />
         <result column="tb_status" jdbcType="INTEGER" property="tbStatus" />
-        <result column="order_date" jdbcType="DATE" property="orderDate" />
+        <result column="order_date"  property="orderDate" />
         <result column="tt_id" jdbcType="INTEGER" property="ttId" />
         <result column="shop_id" jdbcType="VARCHAR" property="shopId" />
         <result column="tt_name" jdbcType="VARCHAR" property="ttName" />
@@ -22,7 +22,24 @@
         -->
         d.tb_id, d.tb_name, d.tb_status, d.order_date, d.tt_id, d.shop_id,t.tt_name,t.tt_capacity
     </sql>
-    <select id="selectByTtId" parameterType="int" resultMap="BaseResultMap">
+
+    <select id="listAll" resultMap="BaseResultMap">
+        SELECT
+        *
+        FROM DinnerTable d
+    </select>
+    <select id="selectCountByTtId" resultType="_int">
+        SELECT
+        COUNT(1)
+        FROM DinnerTable d
+        <where>
+            <if test="_parameter != null and _parameter != '' and _parameter != 0">
+                d.tt_id = #{_parameter,jdbcType=INTEGER}
+            </if>
+        </where>
+    </select>
+
+    <select id="selectByPage"  resultMap="BaseResultMap">
         <!--
           WARNING - @mbg.generated
           This element is automatically generated by MyBatis Generator, do not modify.
@@ -32,10 +49,11 @@
         from DinnerTable d
         LEFT JOIN DinnerTableType t ON d.tt_id=t.tt_id
         <where>
-            <if test="_parameter != null and _parameter != ''">
-                d.tt_id = #{_parameter}
+            <if test="ttId != null and ttId != '' and ttId != 0 ">
+                d.tt_id = #{ttId}
             </if>
         </where>
+        limit #{page.offset},#{page.limit}
     </select>
 
 </mapper>
