+++ backend-manager/backend-manager-dao/src/main/java/cn/yutang/backend/dao/FoodMenuMapper.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="cn.yutang.backend.dao.FoodMenuMapper">
+    <resultMap id="BaseResultMap" type="cn.yutang.backend.pojo.po.Food">
+        <!--
+          WARNING - @mbg.generated
+          This element is automatically generated by MyBatis Generator, do not modify.
+        -->
+        <id column="fd_id" jdbcType="INTEGER" property="fdId" />
+        <result column="fd_name" jdbcType="VARCHAR" property="fdName" />
+        <result column="ft_id" jdbcType="INTEGER" property="ftId" />
+        <result column="fd_price" jdbcType="DOUBLE" property="fdPrice" />
+        <result column="fd_mprice" jdbcType="DOUBLE" property="fdMprice" />
+        <result column="fd_stock" jdbcType="INTEGER" property="fdStock" />
+        <result column="fd_status" jdbcType="INTEGER" property="fdStatus" />
+        <result column="fd_unit" jdbcType="VARCHAR" property="fdUnit" />
+        <result column="fd_remark" jdbcType="VARCHAR" property="fdRemark" />
+        <result column="fd_img" jdbcType="VARCHAR" property="fdImg" />
+        <result column="shop_id" jdbcType="INTEGER" property="shopId" />
+        <result column="fd_recommend" jdbcType="INTEGER" property="fdRecommend" />
+    </resultMap>
+    <sql id="Base_Column_List">
+        <!--
+          WARNING - @mbg.generated
+          This element is automatically generated by MyBatis Generator, do not modify.
+        -->
+        fd_id, fd_name, ft_id, fd_price, fd_mprice, fd_stock, fd_status, fd_unit, fd_remark,
+        fd_img, shop_id, fd_recommend
+    </sql>
+    <select id="selectByPage" resultMap="BaseResultMap">
+        <!--
+          WARNING - @mbg.generated
+          This element is automatically generated by MyBatis Generator, do not modify.
+        -->
+        select
+        <include refid="Base_Column_List" />
+        from Food
+        limit #{offset},#{limit}
+    </select>
+    <select id="selectCount">
+        SELECT COUNT (1)
+        FROM Food
+    </select>
+</mapper>
