+++ backend-manager/backend-manager-interface/src/main/java/cn/yutang/backend/service/IDinnerTableService.java
@@ -1,9 +1,14 @@
 package cn.yutang.backend.service;
 
+import cn.yutang.backend.pojo.dto.MessageResult;
+import cn.yutang.backend.pojo.dto.Page;
 import cn.yutang.backend.pojo.vo.DinnerTableInfo;
 
 import java.util.List;
 
 public interface IDinnerTableService {
-    List<DinnerTableInfo> listTables(Integer ttId);
+    MessageResult<DinnerTableInfo> listTables(Integer ttId, Page page);
+
+    MessageResult<DinnerTableInfo> listTables();
+    List<DinnerTableInfo> allTables();
 }
