+++ backend-manager/backend-manager-interface/src/main/java/cn/yutang/backend/service/IOrderDishesService.java
@@ -5,4 +5,10 @@ import cn.yutang.backend.pojo.po.ShopOrders;
 public interface IOrderDishesService {
 
     void orderDishes(ShopOrders shopOrders);
+
+    int untread(ShopOrders shopOrders);
+
+    int addDishes(ShopOrders shopOrders);
+
+    double payForDishes(ShopOrders orId);
 }
