+++ backend-manager/backend-manager-pojo/backend-manager-pojo.iml
@@ -61,5 +61,10 @@
     <orderEntry type="library" name="Maven: org.mybatis.generator:mybatis-generator-core:1.3.2" level="project" />
     <orderEntry type="library" name="Maven: redis.clients:jedis:2.9.0" level="project" />
     <orderEntry type="library" name="Maven: org.apache.commons:commons-pool2:2.4.2" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.commons:commons-lang3:3.3.2" level="project" />
+    <orderEntry type="library" name="Maven: com.google.zxing:core:3.3.2" level="project" />
+    <orderEntry type="library" name="Maven: aliyun-java-sdk-core:aliyun-java-sdk-core:3.2.2" level="project" />
+    <orderEntry type="library" name="Maven: aliyun-sdk-mns:aliyun-sdk-mns:1.1.8" level="project" />
+    <orderEntry type="library" name="Maven: aliyun-java-sdk-dysmsapi:aliyun-java-sdk-dysmsapi:1.0.0" level="project" />
   </component>
 </module>
