+++ backend-manager/backend-manager-service/src/main/java/cn/yutang/backend/service/impl/FoodOrderServiceImpl.java
@@ -3,12 +3,12 @@ package cn.yutang.backend.service.impl;
 import cn.yutang.backend.dao.FoodOrdersCustomMapper;
 import cn.yutang.backend.pojo.dto.MessageResult;
 import cn.yutang.backend.pojo.dto.Page;
-import cn.yutang.backend.pojo.po.FoodOrders;
 import cn.yutang.backend.pojo.po.FoodOrdersCustom;
 import cn.yutang.backend.pojo.vo.OrderCondition;
 import cn.yutang.backend.service.IFoodOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
 
 import java.util.List;
 
@@ -21,6 +21,7 @@ public class FoodOrderServiceImpl implements IFoodOrderService {
 
     @Override
     public MessageResult<FoodOrdersCustom> findOrders(Page page, OrderCondition orderCondition) {
+
         MessageResult<FoodOrdersCustom> orderResult=new MessageResult<>();
 
         try {
@@ -35,6 +36,37 @@ public class FoodOrderServiceImpl implements IFoodOrderService {
         }
 
         return orderResult;
+
+    }
+
+    @Override
+    public List getPayMethods(Integer shopId) {
+
+        return iFoodOrdersMappers.findMethod(shopId);
+    }
+
+    @Override
+    public List getCustomer(Integer shopId) {
+
+        return iFoodOrdersMappers.findCustomer(shopId);
+    }
+
+    @Override
+    public void getMonthCount(Model model, String date1, String date2) {
+
+        model.addAttribute("MonthCount",iFoodOrdersMappers.getMonthCount(date1,date2));
+    }
+
+    @Override
+    public void getDayCount(Model model, String day1, String day2) {
+
+        Double dayCount = iFoodOrdersMappers.getDayCount(day1, day2);
+
+        if (dayCount==null){
+            dayCount=0.00;
+        }
+
+        model.addAttribute("DayCount",dayCount);
     }
