+++ backend-manager/backend-manager-service/src/main/java/cn/yutang/backend/service/impl/OrderServiceImpl.java
@@ -1,6 +1,8 @@
 package cn.yutang.backend.service.impl;
 
-import cn.yutang.backend.dao.FoodMapper;
+import cn.yutang.backend.dao.FoodMenuMapper;
+import cn.yutang.backend.pojo.dto.MessageResult;
+import cn.yutang.backend.pojo.dto.Page;
 import cn.yutang.backend.pojo.po.Food;
 import cn.yutang.backend.service.IOrderService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,9 +14,14 @@ import java.util.List;
 public class OrderServiceImpl implements IOrderService {
 
     @Autowired
-	FoodMapper foodMapper;
+	FoodMenuMapper foodMenuMapper;
     @Override
-    public List<Food> selectByExample() {
-        return foodMapper.selectByExample(null);
+    public MessageResult<Food> selectByPage(Page page) {
+        MessageResult<Food> messageResult =new MessageResult<>();
+        List<Food> list= foodMenuMapper.selectByPage(page);
+        int count =foodMenuMapper.selectCount();
+        messageResult.setCount(count);
+        messageResult.setData(list);
+        return messageResult;
     }
 }
