+++ backend-manager/backend-manager-web/src/main/webapp/WEB-INF/jsp/pages/food/list.jsp
@@ -20,7 +20,7 @@
     <script src="${pageContext.request.contextPath}/static/js/food/list.js" type="text/javascript"
             charset="utf-8"></script>
 
-    <!--<script type="text/javascript" src="../../static/js/admin.js"></script>-->
+    <script type="text/javascript" src="${pageContext.request.contextPath}/static/js/admin-p.js"></script>
     <!-- 让IE8/9支持媒体查询，从而兼容栅格 -->
     <!--[if lt IE 9]>
     <script src="https://cdn.staticfile.org/html5shiv/r29/html5.min.js"></script>
